update Sveltekit config requirements
Apparently, ``svelte-preprocess`` is not required for UnoCSS. 
I have tested a build without it. Everything works as usual, even the class directives. Just the extractor is enough.

Also, ``target: '#svelte'`` is optional in Sveltekit. Sveltekit uses body of the app if no target is present. 
So I think it isn't necessary to show it in here.