feat: replace fast-glob with fdir
This replaces fast-glob with fdir for file-system traversal.

`fdir` has zero dependencies and is often faster than `fast-glob`.

Note that the `postcss` package is still using `fast-glob` as it has a
requirement for accessing the `Stat` of each found file (which `fdir`
does not expose yet).

If/when fdir exposes the file stats, we can drop `fast-glob` there too.